[Python] Merge units.pyi annotations into units.py
Sixth module of the .pyi stub-merge pass (enhancement #241), and the first to carry
`@overload` method signatures.

- Overloads: UnitSystem.convert_to / convert_activation_energy_to /
  convert_rate_coeff_to each get their three `@overload` stubs prepended to the
  existing implementation. Cython compiles the stubs (the final undecorated `def` is
  the runtime method) and stubtest accepts them; a consumer confirms the published
  return types narrow correctly (`convert_to("3 cm", "m")` -> float,
  `convert_to([...], "m")` -> list[float], `convert_rate_coeff_to(..., UnitStack())`
  -> float). The implementation bodies and their existing arg annotations are left
  unchanged so runtime behavior is identical (this faithfully preserves the prior
  surface, including the stub's `dest: str | Units` on convert_activation_energy_to
  even though the body only handles a string -- a pre-existing discrepancy, not
  something to "fix" in a mechanical merge).
- `new CxxUnitSystem()` -> `make_shared[CxxUnitSystem]()`.
- Typed `__init__` added for Units and UnitSystem (their constructors take args);
  UnitStack needs none (no-arg constructor, default is correct).
- The `_UnitDict` TypedDict is folded inline; `_Array` comes from ._types.
- Allowlist additions for compiled-away names: the `@cython.cfunc` staticmethods
  Units.copy / UnitStack.copy and the cfunc UnitSystem._set_unitSystem, plus the
  cimported _utils helper python_to_anyvalue added to the general allowlist pattern.

cantera.units is added to mypy's ignore_errors list.
