+++ interfaces/cython/.mypyignore
@@ -20,7 +20,7 @@ cantera\.((_|\w)+\.)+__(.*cython|cinit|test|pyx_capi|mutable_keys|readonly_keys|
 # cython`` shim, ``cdef`` helpers cimported from other modules, and ``@cython.cfunc``
 # methods (C-level only). These are expected stub-vs-runtime differences.
 cantera\.\w+\.cython
-cantera\.\w+\.(stringify|pystr|get_from_sparse|make_shared|anymap_to_py|py_to_anymap)
+cantera\.\w+\.(stringify|pystr|get_from_sparse|make_shared|anymap_to_py|py_to_anymap|python_to_anyvalue)
 cantera\.jacobians\.SystemJacobian\.wrap
 cantera\.func1\.func_callback
 cantera\.func1\.Func1\._make_cxx_func1
@@ -29,6 +29,9 @@ cantera\.yamlwriter\.YamlWriter\._get_unitsystem
 cantera\.speciesthermo\.SPECIES_THERMO_\w+
 cantera\.speciesthermo\.wrapSpeciesThermo
 cantera\.speciesthermo\.SpeciesThermo\._assign
+cantera\.units\.Units\.copy
+cantera\.units\.UnitStack\.copy
+cantera\.units\.UnitSystem\._set_unitSystem
 
 # Current output of stubtest --generate-allowlist
 # TODO: Remove many of these after dropping support for mypy 1.19 and older.
